Generalize the ALS cleaner test suite since we now expose the method in RDD
